# cmd/wtd/openapi.json
{
 "components": {
  "responses": {
   "BadRequest": {
    "content": {
     "application/json": {
      "examples": {
       "badName": {
        "value": {
         "error": {
          "code": "invalid_name",
          "message": "name must be 1-64 characters of [A-Za-z0-9._-], must not start with '.', and must not contain '..'"
         }
        }
       }
      },
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    },
    "description": "Validation failure; error.code disambiguates (invalid_name, invalid_path, unknown_project, project_path_missing, path_and_project, bad_request)."
   },
   "Internal": {
    "content": {
     "application/json": {
      "examples": {
       "oops": {
        "value": {
         "error": {
          "code": "internal",
          "message": "unexpected error; see journal"
         }
        }
       }
      },
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    },
    "description": "Unexpected server fault."
   },
...
  },
  "schemas": {
...
   "Meta": {
    "properties": {
     "apiPath": {
      "description": "Prefix every route in this document below `/api/v1` is served under \u2014 append the route's remainder to it, so `/api/v1/sessions` is `apiPath + \"/sessions\"`. Always `/api/v1` today, and the counterpart to `terminalPath`: the two move together if a deployment is ever put behind a path prefix, and reading both is what keeps a client working when they do. Note the scope \u2014 it covers the versioned API only, and says nothing about `/token`, `/healthz`, `/openapi.json`, `/docs/` or `/`, which are not below it and are not described by it.\n",
      "type": "string"
     },
     "features": {
      "description": "Capability strings from the registry in the top-level description. Clients MUST ignore unknown entries and MUST feature-detect anything beyond the ttyd-compatible core before using it \u2014 this list is the only cross-repo skew defense that exists.\n",
      "items": {
       "type": "string"
      },
      "type": "array"
     },
     "hostname": {
      "description": "Short hostname of the machine serving this API.",
      "type": "string"
     },
     "maxSessionNameLength": {
      "description": "The longest session name `POST /api/v1/sessions` will accept on **this** deployment. At most 64, and lower where the server's socket directory is deep enough to leave less room. Validate against this rather than against the schema pattern's 64, and rather than discovering it from a 400.\n",
      "type": "integer"
     },
     "terminalPath": {
      "description": "Path on this server to open the terminal WebSocket at \u2014 combine it with the server's host and port. Always `/ws` today; read it rather than hardcoding, so a future deployment behind a path prefix stays discoverable.\n",
      "type": "string"
     },
     "user": {
      "description": "The service user \u2014 whose shell every terminal hands out.",
      "type": "string"
     },
     "version": {
      "description": "wtd semantic version.",
      "type": "string"
     }
    },
    "required": [
     "version",
     "features",
     "hostname",
     "user",
     "terminalPath",
     "apiPath",
     "maxSessionNameLength"
    ],
    "type": "object"
   },
...
   "Project": {
    "properties": {
     "exists": {
      "description": "Whether `path` is currently a directory. False means a `POST` naming this project is refused with `project_path_missing` (400).\n",
      "type": "boolean"
     },
     "name": {
      "type": "string"
     },
     "path": {
      "type": "string"
     }
    },
    "required": [
     "name",
     "path",
     "exists"
    ],
    "type": "object"
   },
...
  }
 },
...
 "info": {
...
  "license": {
   "identifier": "MIT",
   "name": "MIT"
  },
  "summary": "Browser terminal daemon: ttyd-wire-compatible WebSocket terminal plus a JSON API over dtach sessions.\n",
  "title": "wtd HTTP API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/meta": {
   "get": {
    "description": "The anti-skew endpoint (see the features registry in the top-level description). Call\nit once each time you connect to a server and cache it no longer than that: it is cheap,\nand the whole point is noticing a server that changed underneath an installed client.\nCaching across app launches defeats the purpose.\n",
    "operationId": "getMeta",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "examples": {
         "devbox": {
          "value": {
           "apiPath": "/api/v1",
           "features": [
            "token",
            "url-arg",
            "writable",
            "sessions-api",
            "projects-api",
            "picker-ui",
            "scrollback-replay",
            "docs-endpoint",
            "session-read",
            "session-status",
            "session-narration"
           ],
           "hostname": "devbox",
           "maxSessionNameLength": 64,
           "terminalPath": "/ws",
           "user": "sam",
           "version": "0.2.0"
          }
         }
        },
        "schema": {
         "$ref": "#/components/schemas/Meta"
        }
       }
      },
      "description": "Server metadata."
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "Server identity and capability list"
   }
  },
  "/api/v1/projects": {
   "get": {
    "description": "The shortcuts this deployment has configured, already parsed \u2014 a client never\nreads the file itself. Each entry reports `exists`, whether its path is\ncurrently a directory, so a client can warn before creating a session\nsomewhere unexpected. A deployment with no shortcuts configured, or an\nunreadable file, is an empty list rather than an error.\n",
    "operationId": "listProjects",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "examples": {
         "two": {
          "value": [
           {
            "exists": true,
            "name": "dev",
            "path": "/home/sam/src/my-app"
           },
           {
            "exists": false,
            "name": "ops",
            "path": "/home/sam/src/infra"
           }
          ]
         }
        },
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Project"
         },
         "type": "array"
        }
       }
      },
      "description": "All shortcuts (possibly empty array).",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-store",
         "type": "string"
        }
       }
      }
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "List project shortcuts"
   }
  },
  "/api/v1/sessions": {
   "get": {
    "description": "Every session this server can see, sorted ascending by the byte value of `name`. Sessions\ncan be created by deep link as well as by `POST`, and every surface that lists them must\nagree about what exists \u2014 which drives both rules below.\n\n1. **Looser names than this API would create are still listed.** A session made by\n   deep-linking `/ws?arg=my%20proj` is named `my proj`; it appears here even though\n   `POST` would reject that name. Do not assume listed names satisfy the create-side\n   pattern. Names beginning with `.` are listed too, even though `POST` refuses them.\n2. **Dead sessions are not listed, and this request removes them.** A socket\n   left behind by a session that is gone would otherwise show as a session\n   that cannot be attached, so this endpoint verifies each one and unlinks the\n   ones it can prove are dead. That probing is also why this read is not instant:\n   budget a client timeout of a few seconds rather than a few hundred milliseconds, and\n   expect the cost to grow with the number of sessions. Yes, that means a `GET` has a\n   side effect; it is\n   the only read every client already makes, which is why it is the one that\n   converges what exists with what is listed. Why, and the guard against racing a\n   session being recreated under the same name: `/docs/session-lifecycle.md` \u00a75.\n",
    "operationId": "listSessions",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "examples": {
         "two": {
          "value": [
           {
            "attached": false,
            "attachedCount": 0,
            "createdAt": "2026-07-22T21:03:00Z",
            "cwd": "/home/sam/src/infra",
            "name": "ops",
            "pid": 51234
           },
           {
            "attached": true,
            "attachedCount": 2,
            "createdAt": "2026-07-24T14:09:00Z",
            "cwd": "/home/sam/src/riggs",
            "name": "riggs",
            "pid": 60211
           }
          ]
         }
        },
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Session"
         },
         "type": "array"
        }
       }
      },
      "description": "All live sessions (possibly empty array).",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-store",
         "type": "string"
        }
       }
      }
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "List dtach sessions"
   },
   "post": {
    "description": "Creates a dtach session **without attaching**, for naming a working directory or\nproject up front. It is **not** a precondition for deep-linking: `/ws?arg=<name>`\ncreates the session if it does not exist, so a client can open one directly and never\ncall this.\n\nA session created here is **indistinguishable** from one created by deep link \u2014 same\ninvocation, same environment, built from the same code. That parity is mandatory, so\nanything you can do with a deep-linked session you can do with this one, and nothing\nabout it will surprise a user who switches between the two.\nThe invocation and the reasons each part of it matters: `/docs/session-lifecycle.md` \u00a76.\n\nWorking directory: `project` (a shortcut name from `GET /api/v1/projects`) or\n`path` (absolute, must exist), else the service user's `$HOME`. The two are\nmutually exclusive. Where a deep-linked connection silently falls back to `$HOME`\nfor a directory that is missing, this API returns an error instead \u2014 a program\nshould be told, not guessed for.\n\nName rules are **stricter than the deep-link path's**, deliberately: no leading `.`\nand no `..` anywhere, the latter because such a session could never be reached by\ndeep link. See the `name` property for the length ceiling, which is\ndeployment-dependent.\n\nIf the name belongs to a session that is already gone, its leftover socket is\nremoved and creation proceeds \u2014 so retrying a create for a session that has\ndied does the obvious thing rather than returning a conflict.\n",
    "operationId": "createSession",
    "requestBody": {
     "content": {
      "application/json": {
       "examples": {
        "bare": {
         "value": {
          "name": "scratch"
         }
        },
        "byPath": {
         "value": {
          "name": "riggs",
          "path": "/home/sam/src/riggs"
         }
        },
        "byProject": {
         "value": {
          "name": "ops",
          "project": "ops"
         }
        }
       },
       "schema": {
        "$ref": "#/components/schemas/SessionCreateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "examples": {
         "created": {
          "value": {
           "attached": false,
           "attachedCount": 0,
           "createdAt": "2026-07-24T21:00:00Z",
           "cwd": "/home/sam/src/riggs",
           "name": "riggs",
           "pid": 61402
          }
         }
        },
        "schema": {
         "$ref": "#/components/schemas/Session"
        }
       }
      },
      "description": "Session created, detached.",
      "headers": {
       "Location": {
        "description": "Canonical resource path for the new session.",
        "schema": {
         "examples": [
          "/api/v1/sessions/riggs"
         ],
         "type": "string"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/BadRequest"
     },
     "403": {
      "$ref": "#/components/responses/OriginForbidden"
     },
     "409": {
      "content": {
       "application/json": {
        "examples": {
         "exists": {
          "value": {
           "error": {
            "code": "already_exists",
            "message": "session 'riggs' already exists"
           }
          }
         }
        },
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "A session with this name already exists. Two concurrent creates of the same name resolve to one winner and one 409. A session whose death cannot be established counts as existing, and is reported rather than silently replaced."
     },
     "413": {
      "$ref": "#/components/responses/PayloadTooLarge"
     },
     "415": {
      "$ref": "#/components/responses/UnsupportedMediaType"
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "Create a detached session"
   }
  },
  "/api/v1/sessions/{name}": {
   "delete": {
    "description": "Terminates the session and removes its socket; `204` once the socket is gone.\n\n**No path is ever constructed from `{name}`** \u2014 it is URL-decoded and matched\nbyte-exact against the names that actually exist. That is why this parameter\nhas no validation pattern: a deep link creates names `POST` would reject, and those\nsessions must be deletable too. A name matching nothing is `404`, including anything\ncontaining `/`.\n\nDeleting an **attached** session is allowed. The live shell is terminated under\nwhoever is watching it, and their connection closes normally rather than erroring;\nthis API does not guess whether that is what you meant.\nDeleting a session that is **already gone** also succeeds \u2192 `204`, including one\nthat has become unreachable by every other route, so `DELETE` is the way to\nclean up after a session nothing else can address.\n\nThe session's background jobs are terminated along with it. Termination is\nbounded at roughly **7 s** worst case \u2014 budget a client timeout above that \u2014\nwhich is why this endpoint answers synchronously instead of handing back a job\nto poll. Ordering, escalation, and the phantom-session failure modes the\nordering exists to prevent: `/docs/session-lifecycle.md` \u00a77.\n",
    "operationId": "deleteSession",
    "parameters": [
     {
      "description": "Exact session name as returned by a list response, URL-encoded as needed.\n",
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Session terminated and its socket removed."
     },
     "403": {
      "$ref": "#/components/responses/OriginForbidden"
     },
     "404": {
      "content": {
       "application/json": {
        "examples": {
         "missing": {
          "value": {
           "error": {
            "code": "not_found",
            "message": "no session named 'riggs'"
           }
          }
         }
        },
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "No socket with this name exists."
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "Termination did not complete inside the bounded ~7 s. The session is left as it was and is still listed, so a retry is safe."
     }
    },
    "summary": "Terminate a session and clean up its socket"
   },
   "get": {
    "description": "The same object the list returns, for one name \u2014 this is what the `Location` header on a `201` points at. `{name}` is URL-decoded and matched byte-exact against the names that exist, so a name the list shows is always fetchable even when it would fail the create-side rules. Detect with the `session-read` feature; a server without it forces a list-and-filter instead.\n",
    "operationId": "getSession",
    "parameters": [
     {
      "description": "Exact session name as returned by a list response, URL-encoded as needed.",
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Session"
        }
       }
      },
      "description": "The session.",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-store",
         "type": "string"
        }
       }
      }
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "No session with this name exists."
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "Read one session"
   }
  },
  "/api/v1/sessions/{name}/narration": {
   "get": {
    "description": "A short summary of the session's last turn, written to be **spoken aloud** rather than\nread. Intended for a hands-free client: a voice that says what an agent just did, and\nsays when it is blocked on you.\n\nThis is not the session's output. Terminal output is raw bytes with cursor addressing in\nit, and an agent that redraws its interface produces a repaint rather than a sentence \u2014\nso a client that wants something speakable must ask for this instead of transforming\n`/ws` output, which `/docs/ws-protocol.md` \u00a76 forbids anyway.\n\n**`404` is the normal answer and means \"nothing to say\".** Three different situations\nproduce it and none is an error: this deployment has no narration configured, the agent\nin that session does not report, or it has not finished a turn yet. A polling client\ntreats `404` as silence and keeps polling.\n\n`{name}` is URL-decoded and matched byte-exact against the names that exist, before any\npath is built from it \u2014 the same rule `getSession` and `deleteSession` follow. Detect with\nthe `session-narration` feature.\n\n## What a client must do\n\nSpeak a summary **at most once**: compare `at` against the last value you spoke and\nignore anything not newer. Nothing on the server tracks what a client has said, and a\nclient that re-speaks on every poll narrates the same sentence every few seconds.\n\nSpeak **unprompted only when `needsYou` is `true`**. That is the case where the session\nis blocked and cannot continue without an answer. Everything else is available when the\nuser asks for it and should not interrupt them.\n\n`event` uses the vocabulary of `agentStatus`. Treat a value you do not recognize as\n\"some event\", never as none \u2014 the same forward-compatibility rule \u00a76a puts on `WTState`.\n",
    "operationId": "getSessionNarration",
    "parameters": [
     {
      "description": "Exact session name as returned by a list response, URL-encoded as needed.",
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Narration"
        }
       }
      },
      "description": "The summary.",
      "headers": {
       "Cache-Control": {
        "schema": {
         "const": "no-store",
         "type": "string"
        }
       }
      }
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": "No session with this name exists, or it has no summary. Not an error; see the description.\n"
     },
     "500": {
      "$ref": "#/components/responses/Internal"
     }
    },
    "summary": "Read the spoken-form summary of what a session last did"
   }
  },
...
 },
 "servers": [
  {
   "description": "wtd's single listener, and the only server that serves this document. host is the resolved bind address \u2014 a tailnet IP, 127.0.0.1, or an interface address, never a wildcard.\n**Use port 7681 (`WT_PORT`).** One box runs one listener on one port, and it serves every route in this document. A server that answers `/`, `/token` and `/ws` but returns 404 for `/openapi.json`, the `/api/v1/*` routes and the picker is an older deployment that has not been upgraded; `GET /api/v1/meta` succeeding is the cheapest proof you are talking to wtd, and its `features[]` is how to test for anything newer than this revision.\n",
   "url": "http://{host}:{port}",
   "variables": {
    "host": {
     "default": "127.0.0.1"
    },
    "port": {
     "default": "7681"
    }
   }
  }
 ]
}